chore(BUX-200): move logging methods to separate package
